lisp: initial commit
% go doc
Lisp is an implementation of the language defined, with sublime concision,
in the first few pages of the LISP 1.5 Programmer's Manual by McCarthy,
Abrahams, Edwards, Hart, and Levin, from MIT in 1962.

It is a pedagogical experiment to see just how well the interpreter
(actually EVALQOUTE/APPLY) defined on page 13 of that book really works. The
answer is: perfectly, of course.

This program's purpose was fun and education, and in no way to create a
modern or even realistic Lisp implementation. The goal was to turn that
marvelous page 13 into a working interpreter using clean, direct Go code.

The program therefore has several profound shortcomings, even with respect
to the Lisp 1.5 book:

    - No `SET` or `SETQ`.
    - No `PROG`. The intepreter, by calling `APPLY`, can evaluate only a single
      expression, a possibly recursive function invocation. But this is Lisp, and
      that's still a lot.
    - No character handling.
    - No I/O. Interactive only, although it can start by reading a file specified
      on the command line.

It is slow and of course the language is very, very far from Common Lisp or
Scheme.
%